requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
Markdown==3.3.3
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
